@@ -1,24 +1,25 @@
 /* eslint-disable no-console */
 /* eslint-disable global-require */
 /* eslint-disable @typescript-eslint/no-var-requires */
-const path = require('path')
-const isDev = require('electron-is-dev')
-const { app, nativeTheme, BrowserWindow, ipcMain, shell } = require('electron')
-const { isCC } = require('./app/core.js')
-const { createWindow } = require('./app/win.js')
-const { createTray } = require('./app/tray.js')
-const { startInstance, closeAllSubs } = require('./app/instances.js')
-const { handlers } = require('./app/handlers.js')
-const { installDevtools } = require('./app/devtools.js')
-const { setupProtocol, handleProtocol } = require('./app/protocol.js')
+import fs from 'fs'
+import path from 'path'
+import isCC from './app/utils/isCC.mjs'
+import isDev from 'electron-is-dev'
+import createWindow from './app/utils/win.mjs'
+import { app, nativeTheme, BrowserWindow, ipcMain, shell, session } from 'electron'
+import { createTray } from './app/utils/tray.mjs'
+import { startInstance, closeAllSubs } from './app/instances.js'
+import { handlers } from './app/handlers.js'
+import { setupProtocol, handleProtocol } from './app/protocol.js'
+import { EventEmitter } from 'events'
+import { fileURLToPath } from 'node:url'
 
-require('events').EventEmitter.defaultMaxListeners = 15
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
-let installExtension
-if (isDev) {
-  const devTools = require('electron-devtools-installer')
-  installExtension = devTools.default
-}
+EventEmitter.defaultMaxListeners = 15
+
+const reduxDevtoolsPath = 'C:\\Users\\49152\\AppData\\Local\\Google\\Chrome\\User Data\\Default\\Extensions\\lmhkpmbekcpmknklioeibfkpmmfibljd\\3.2.7_0'
 
 const subpy = null
 const subprocesses = {}
@@ -28,8 +29,16 @@ let win
 setupProtocol()
 const gotTheLock = app.requestSingleInstanceLock()
 
+if (!fs.existsSync(path.join(app.getPath('userData'), '.ledfx-cc'))) {
+  console.log('Creating .ledfx-cc folder')
+  fs.mkdirSync(path.join(app.getPath('userData'), '.ledfx-cc'))
+}
+
 const ready = () =>
   app.whenReady().then(async () => {
+    if (isDev) {
+      await session.defaultSession.loadExtension(reduxDevtoolsPath)
+    }
     nativeTheme.themeSource = 'dark'
     const thePath = process.env.PORTABLE_EXECUTABLE_DIR || path.resolve('.')
 
@@ -37,7 +46,8 @@ const ready = () =>
       ? createWindow(win, { additionalArguments: ['integratedCore'] })
       : createWindow(win)
 
-    require('@electron/remote/main').enable(wind.webContents)
+    const remoteMain = await import('@electron/remote/main/index.js')
+    remoteMain.enable(wind.webContents)
 
     wind.webContents.setWindowOpenHandler(({ url }) => {
       if (url.includes(' https://accounts.spotify.com/authorize')
@@ -49,8 +59,8 @@ const ready = () =>
       return { action: 'allow' }
     })
 
-    if (isCC) startInstance(wind, 'instance1', subprocesses)
-    if (isDev) installDevtools(installExtension)
+    if (isCC) startInstance(wind, 'instance1', subprocesses)      
+   
 
     createTray(isCC, wind, thePath, __dirname)
 
